Change import style for Group.py into Molecule.py to avoid circular imports
Also added the import into Group.py and the start of a function to show that the circular import error is not present.

If an ImportError or AttributeError starts appearing in the future, a topic I read on stackoverflow suggests using only the absolute import (i.e. import rmgpy.molecule.molcule without the 'as' statement). The reason I don't want to use this right now is because it makes any object calls really long.