Correcting updateLonePairs
The calculation of lone pairs was wrong for the N1s atom type,
and could potentially be worng for some of the new sulfur atom types
which will soon be introduced in RMG.

This commit fixes the calculation of the number of lone pairs by using
the `valence_electrons` attribute of an element. An error will be printed
incase the number of lone pairs isn't an integer.

The previous calculation resulted in 3 lone pairs for N1s (where it should only have 2):
atom1.lonePairs = 4 - atom1.radicalElectrons - int(order) = 4 - 1 = 3
The new calculation will give (5-1)/2 = 2 for N1s.
The new calculation will collapse to be identical for carbon \ oxygen \ silicon.